+++ R/get_filename.R
@@ -1,7 +1,3 @@
-get_filename <- function(file_path){
-        
-        f_sep <- .Platform$file.sep
-        split_path <- unlist(strsplit(file_path,f_sep))
-        file_name <- unlist(strsplit(split_path[length(split_path)],"[.]"))[1]
-        return(file_name)
+get_filename <- function(file_path) {
+        tools::file_path_sans_ext(basename(file_path))
 }