+++ tests/testthat/test-cds2aa.R
@@ -1,1 +1,0 @@
-context("Test: cds2aa()")