+++ tests/testthat/test-extract_features.R
@@ -1,1 +1,0 @@
-context("Test: extract_features()")