+++ tests/testthat/test-generate_ortholog_table.R
@@ -1,1 +1,0 @@
-context("Test: generate_ortholog_tables()")