+++ tests/testthat/test-map_generator_lnc.R
@@ -1,1 +1,0 @@
-context("Test: map_generator_lnc()")