+++ tests/testthat/test-orthologs_lnc.R
@@ -1,1 +1,0 @@
-context("Test: orthologs_lnc()")