+++ tests/testthat/test-set_blast.R
@@ -1,1 +1,0 @@
-context("Test: set_blast()")