+++ NEWS.md
@@ -1,3 +1,16 @@
+## `orthologr` version 0.4.3
+
+### New Function
+
+- `deepclust()` enables `diamond deepclust` functionalities for clustering protein sequences. Multiple input `.fasta` files or `.fasta.gz` are supported.
+- `deepclust_realign()` enables `diamond deepclust` functionalities for realigning protein sequences from the output of `deepclust()`. Multiple input `.fasta` files or `.fasta.gz` are supported.
+- `deepclust_annotate()` enables `diamond deepclust` functionalities for annotating protein sequences from the output of `deepclust()` based on user provided names for each proteome file (or the name of the proteome file if not provided). `deepclust_annotate()` can also be used to annotate the output of `deepclust_realign()`. Multiple input `.fasta` files or `.fasta.gz` are supported.
+
+### Miscellaneous
+
+- Unit tests for `deepclust()`, `deepclust_realign()` and `deepclust_annotate()` have been added to ensure the correct functionality of the new features.
+- Unit tests for `dNdS()`, `divergence_stratigraphy()` and `divergence_map()` functions have been created.
+
 ## `orthologr` version 0.4.2
 
 ### New Function